requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements.txt requirements.in
#
afinn==0.1
amqp==2.3.2               # via kombu
aniso8601==7.0.0          # via flask-restful
appnope==0.1.0            # via ipython
astroid==1.6.5            # via pylint
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
billiard==3.5.0.4         # via celery
celery==4.2.0
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via dynaconf, flask
decorator==4.3.0          # via ipython, traitlets
dynaconf==1.0.2
elasticsearch==6.3.0
fanstatic==1.0.0          # via js
flake8==3.5.0
flask-injector==0.12.0
flask-restful==0.3.7
flask==1.0.2
funcy==1.10.3
greenlet==0.4.14          # via meinheld
gunicorn==19.9.0
idna==2.7                 # via requests
injector==0.17.0          # via flask-injector
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==5.7.0            # via ipdb
isort==4.3.4              # via pylint
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
js==1.0
kombu==4.2.1              # via celery
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8, pylint
meinheld==0.6.1
mock==3.0.5
more-itertools==4.2.0     # via pytest
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pluggy==0.6.0             # via pytest
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.5.4                 # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via ipython
pylint==1.9.2
pytest-flask==0.10.0
pytest-html==1.19.0
pytest-metadata==1.7.0    # via pytest-html
pytest==3.6.3
python-box==3.2.0         # via dynaconf
python-dotenv==0.8.2      # via dynaconf
pytz==2018.5              # via celery, flask-restful
redis==2.10.6
requests-mock==1.5.0
requests==2.19.1
selenium==3.13.0
shutilwhich==1.1.0        # via fanstatic
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via astroid, dynaconf, flask-restful, mock, more-itertools, prompt-toolkit, pylint, pytest, requests-mock, traitlets
timeout-decorator==0.4.1
toml==0.9.4               # via dynaconf
traitlets==4.3.2          # via ipython
typing==3.7.4.1           # via flask-injector
urllib3==1.23             # via elasticsearch, requests
vine==1.1.4               # via amqp
wcwidth==0.1.7            # via prompt-toolkit
webob==1.8.2              # via fanstatic
werkzeug==0.14.1          # via flask, pytest-flask
wrapt==1.10.11            # via astroid
